# f1_project/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "F1 Weekend Intelligence",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/events/{event_id}": {
      "get": {
        "summary": "Overview",
        "operationId": "overview_api_v1_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}/laps": {
      "get": {
        "summary": "Laps",
        "operationId": "laps_api_v1_events__event_id__laps_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "drivers",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Drivers"
            }
          },
          {
            "name": "session",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "FP1",
                    "FP2",
                    "FP3"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          },
          {
            "name": "compound",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compound"
            }
          },
          {
            "name": "usable",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Usable"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "Time",
                "LapTime",
                "LapNumber",
                "Driver",
                "Sector1Time",
                "Sector2Time",
                "Sector3Time",
                "SpeedST",
                "TyreLife"
              ],
              "type": "string",
              "default": "Time",
              "title": "Sort"
            }
          },
          {
            "name": "descending",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Descending"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 4000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LapPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}/compare": {
      "get": {
        "summary": "Comparison",
        "operationId": "comparison_api_v1_events__event_id__compare_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "drivers",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Drivers"
            }
          },
          {
            "name": "session",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "FP1",
                    "FP2",
                    "FP3"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          },
          {
            "name": "compound",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compound"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comparison"
                  },
                  "title": "Response Comparison Api V1 Events  Event Id  Compare Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/quality": {
      "get": {
        "summary": "Quality",
        "operationId": "quality_api_v1_quality_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quality"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}/predictions": {
      "get": {
        "summary": "Predictions",
        "operationId": "predictions_api_v1_events__event_id__predictions_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PredictionRow"
                  },
                  "title": "Response Predictions Api V1 Events  Event Id  Predictions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predict": {
      "post": {
        "summary": "What If",
        "operationId": "what_if_api_v1_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predict/custom": {
      "post": {
        "summary": "Predict Custom",
        "operationId": "predict_custom_api_v1_predict_custom_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}/export": {
      "get": {
        "summary": "Export",
        "operationId": "export_api_v1_events__event_id__export_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "laps",
                "results",
                "predictions",
                "audit"
              ],
              "type": "string",
              "default": "laps",
              "title": "Kind"
            }
          },
          {
            "name": "drivers",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Drivers"
            }
          },
          {
            "name": "session",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "FP1",
                    "FP2",
                    "FP3"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          },
          {
            "name": "compound",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compound"
            }
          },
          {
            "name": "usable",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Usable"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "Time",
                "LapTime",
                "LapNumber",
                "Driver",
                "Sector1Time",
                "Sector2Time",
                "Sector3Time",
                "SpeedST",
                "TyreLife"
              ],
              "type": "string",
              "default": "Time",
              "title": "Sort"
            }
          },
          {
            "name": "descending",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Descending"
            }
          },
          {
            "name": "result_sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "Position",
                "Driver",
                "Q1",
                "Q2",
                "Q3"
              ],
              "type": "string",
              "default": "Position",
              "title": "Result Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Comparison": {
        "properties": {
          "driver": {
            "type": "string",
            "title": "Driver"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "median": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median"
          },
          "iqr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iqr"
          },
          "enough_laps": {
            "type": "boolean",
            "title": "Enough Laps"
          },
          "best_lap": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Lap"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "driver",
          "count",
          "median",
          "iqr",
          "enough_laps",
          "best_lap"
        ],
        "title": "Comparison"
      },
      "CustomRequest": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "FP1": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PracticeInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "FP2": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PracticeInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "FP3": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PracticeInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id"
        ],
        "title": "CustomRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LapPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Lap"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "LapPage"
      },
...
      "Overview": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/Event"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/Result"
            },
            "type": "array",
            "title": "Results"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/Session"
            },
            "type": "array",
            "title": "Sessions"
          },
          "weather": {
            "items": {
              "$ref": "#/components/schemas/Weather"
            },
            "type": "array",
            "title": "Weather"
          }
        },
        "type": "object",
        "required": [
          "event",
          "results",
          "sessions",
          "weather"
        ],
        "title": "Overview"
      },
...
      "PracticeInput": {
        "properties": {
          "time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ],
            "title": "Time"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "time"
        ],
        "title": "PracticeInput"
      },
...
      "Quality": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "checksums": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Checksums"
          },
          "raw_laps": {
            "type": "integer",
            "title": "Raw Laps"
          },
          "raw_results": {
            "type": "integer",
            "title": "Raw Results"
          },
          "kept_laps": {
            "type": "integer",
            "title": "Kept Laps"
          },
          "removals": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Removals"
          },
          "overlapping_flags": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Overlapping Flags"
          },
          "after_qualifying_laps": {
            "type": "integer",
            "title": "After Qualifying Laps"
          },
          "pre_qualifying_laps": {
            "type": "integer",
            "title": "Pre Qualifying Laps"
          },
          "result_duplicates": {
            "type": "integer",
            "title": "Result Duplicates"
          },
          "missing_targets": {
            "type": "integer",
            "title": "Missing Targets"
          },
          "no_practice_rows": {
            "type": "integer",
            "title": "No Practice Rows"
          },
          "feature_rows": {
            "type": "integer",
            "title": "Feature Rows"
          },
          "missing_features": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Missing Features"
          }
        },
        "type": "object",
        "required": [
          "version",
          "checksums",
          "raw_laps",
          "raw_results",
          "kept_laps",
          "removals",
          "overlapping_flags",
          "after_qualifying_laps",
          "pre_qualifying_laps",
          "result_duplicates",
          "missing_targets",
          "no_practice_rows",
          "feature_rows",
          "missing_features"
        ],
        "title": "Quality"
      },
...
      "ScenarioRequest": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "driver": {
            "type": "string",
            "title": "Driver"
          },
          "overrides": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "FP1_Time",
                "FP2_Time",
                "FP3_Time"
              ]
            },
            "type": "object",
            "title": "Overrides"
          },
          "sessions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/PracticeInput"
            },
            "propertyNames": {
              "enum": [
                "FP1",
                "FP2",
                "FP3"
              ]
            },
            "type": "object",
            "title": "Sessions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "driver"
        ],
        "title": "ScenarioRequest"
      },
...
    }
  }
}